move docker func
