+++ edgelet/iotedge/src/check/checks/mod.rs
@@ -37,3 +37,42 @@ pub use self::storage_mounted_from_host::{EdgeAgentStorageMounted, EdgeHubStorag
 pub use self::well_formed_config::WellFormedConfig;
 pub use self::well_formed_connection_string::WellFormedConnectionString;
 pub use self::windows_host_version::WindowsHostVersion;
+
+use std::ffi::OsStr;
+use std::process::Command;
+
+use failure::{self, Context, Fail};
+
+pub(crate) fn docker<I>(
+    docker_host_arg: &str,
+    args: I,
+) -> Result<Vec<u8>, (Option<String>, failure::Error)>
+where
+    I: IntoIterator,
+    <I as IntoIterator>::Item: AsRef<OsStr>,
+{
+    let mut process = Command::new("docker");
+    process.arg("-H");
+    process.arg(docker_host_arg);
+
+    process.args(args);
+
+    let output = process.output().map_err(|err| {
+        (
+            None,
+            err.context(format!("could not run {:?}", process)).into(),
+        )
+    })?;
+
+    if !output.status.success() {
+        let stderr = String::from_utf8_lossy(&*output.stderr).into_owned();
+        let err = Context::new(format!(
+            "docker returned {}, stderr = {}",
+            output.status, stderr,
+        ))
+        .into();
+        return Err((Some(stderr), err));
+    }
+
+    Ok(output.stdout)
+}
